requirements.txt
Django==4.0.5
django_allauth==0.51.0
django-crispy-forms==1.13.0
requests==2.25.1